requirements.txt
fastapi==0.62.0
uvicorn==0.13.1
opencv-python-headless
numpy==1.18.2
Pillow==8.2.0
tensorflow==2.3.1
python-decouple